**IMPORTANT: Please do not create a Pull Request without creating an
issue first.**

## Summary

**Issue:** Enter the issue(s) this pull request resolves.

A clear and concise description of this pull request

## Type (Fill in "x" to check)

- [ ] Bug Fix
- [ ] New Feature

## Checklist

Please make sure that all items are checked before submitting this
request.

- [ ] Code linter has been run and issues have all been resolved
- [ ] The code has been thoroughly tested and no visible bugs have been
introduced
- [ ] The pull request will completely resolve the issue(s) mentioned
- [ ] The pull request only resolves the issue(s) mentioned and nothing
more

## Additional context

Add any other context here.

+++ Taipei-City-Dashboard-DE/dags/proj_city_dashboard/energy_saving_coaching_situation_tpe_gov/energy_saving_coaching_situation_tpe_gov.py
@@ -63,9 +63,9 @@ def _transfer(**kwargs):
     # Transfer year from ROC to AD
     data['year'] = data['year'].astype(int) + 1911
     data['compliance_rate'] = data['compliance_rate'].replace('%', '', regex=True)
-    data['compliance_rate'] = data['compliance_rate'].astype(int)
+    data['compliance_rate'] = data['compliance_rate'].astype(float)
     data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].replace('%', '', regex=True)
-    data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].astype(int)
+    data['reinspection_compliance_rate'] = data['reinspection_compliance_rate'].astype(float)
     data['audited_households'] = data['audited_households'].astype(int)
     data['carbon_sink_daan_forest_parks'] = data['carbon_sink_daan_forest_parks'].astype(int)
     data['power_saving_ten_thousand_kWh_per_year'] = data['power_saving_ten_thousand_kWh_per_year'].astype(int)
